Materials/Lecture_3/requirements.txt
numpy
opencv-python
faiss-gpu
fastapi
uvicorn